# Performance improvement on importing