--- a/word_forms/constants.py
+++ b/word_forms/constants.py
@@ -1,20 +1,6 @@
-from difflib import get_close_matches
 from collections import defaultdict
 from pathlib import Path
-
-try:
-    from nltk.corpus import wordnet as wn
-    raise_lookuperror_if_wordnet_data_absent = wn.synsets("python")
-except LookupError:
-    import nltk
-    nltk.download("wordnet")
-try:
-    from nltk.corpus import words
-except LookupError:
-    nltk.download("words")
-
-ALL_WORDNET_WORDS = set(words.words())
-
+import json
 
 class Verb(object):
     def __init__(self, verbs=None):
@@ -38,23 +24,6 @@ for line in lines:
         for verb in verbs:
             CONJUGATED_VERB_DICT[verb].update(verbs)
 
-ADJECTIVE_TO_ADVERB = {"good" : "well", "fast" : "fast", "hard" : "hard",
-                       "late" : "late", "early" : "early", "daily" : "daily",
-                       "straight" : "straight"}
-for ss in wn.all_synsets(pos="r"):
-    for lemma in ss.lemmas():
-        word = lemma.name()
-        this_word_lemmas = [
-            lemma
-            for ss in wn.synsets(word, pos=wn.ADV)
-            for lemma in ss.lemmas()
-            if lemma.name() == word
-        ]
-        pertainyms = {
-            pertainym.name()
-            for this_word_lemma in this_word_lemmas
-            for pertainym in this_word_lemma.pertainyms()
-        }
-        matches = get_close_matches(word, pertainyms)
-        for match in matches:
-            ADJECTIVE_TO_ADVERB[match] = word
+adj_fh = open(str(Path(__file__).parent.absolute() / "adj_to_adv.txt"))
+ADJECTIVE_TO_ADVERB = json.load(adj_fh)
+adj_fh.close()