… word_forms/adj_to_adv.txt