Closes #20 

Hello!

---
## Pull request overview
This PR is about a performance improvement with three main changes:
1. Removing the now-defunct `ALL_WORDNET_WORDS` as inspired by #20.
2. Making `ADJECTIVE_TO_ADVERB` be loaded from a file rather than through NLTK's Wordnet.
3. (Slightly) changing how Wordnet is imported in `word_forms.py`.

---

### Change 1:
Before my commit 359fc0c8dba0a2d7308f2ffd3f0dc5f6b37025f0, there existed a function `singularize(noun)`, which attempted to return the singular form of a noun, and checked if it existed in Wordnet using `singular in ALL_WORDNET_WORDS`. Essentially, it tried to find a form of the word passed to `get_word_forms` so the rest of the program could work with a word in Wordnet.
In this commit I change this function to using `WordNetLemmatizer().lemmatize()`, which as property has that it ensures that the passed word is in Wordnet. However, the import of `ALL_WORDNET_WORDS` remained, without needing to exist.
In @gutfeeling 's commit 143abc35b01971f6d23ad63f826c3e1f27e15225 this import was removed, but the initialisation in `constants.py` remained unnecessarily.

### Change 2:
I tried to look for more ways to improve import times, as they were roughly 2.8 seconds for me earlier. I looked at `constants`, and realised that the dict that is created for `ADJECTIVE_TO_ADVERB` is only ~2800 items big, while it costs me just under a second to compute. Because this dict is unlikely to change much between different runs of the program (it can only change if NLTK's Wordnet is updated), this second can be gained by simply moving this dict into a file, and loading that - very similarly to what is done with `en-verbs.txt`.

So, with change 1 and 2 combined, there is no use for NLTK's wordnet in `constants.py`, so those imports are removed too.

## Change 3:
As you can see in the list of commits, I also attempted a faster method of checking whether NLTK's wordnet is downloaded:
```python
try:
    from nltk.data import find
    find("corpora/wordnet.zip")
except LookupError:
    from nltk import download
    download("wordnet", quiet=True)
from nltk.corpus import wordnet as wn
```
This is considerably faster for importing (improving import times fourfold), but will increase the execution time of the very first call of `get_word_forms()` by roughly 1.5 seconds.
This is because NLTK lazy loads its wordnet data, and the original check that uses `wn.synsets("python")` will cause NLTK to load wordnet, costing 1.5 seconds.

So, the question is: Would you rather wait 1.5 seconds at the start of your program, or 1.5 extra seconds whenever you first call `get_word_forms`, and my take is that any program that calls the function in a "live" environment, e.g. based on user input, will rather have the former. So, I didn't implement this change, and only changed `wn.synsets("python")` to `wn.synset("python.n.01")`, which simply gets one synset instead of 3. This is marginably faster.

---

### Rough performance changes
The import time of this module has been reduced by about 35%, from ~2.3975 to ~1.5668s (median time between 5 runs each), with no performance difference of the `get_word_forms` function. Furthermore, all tests still run perfectly.

---

Thank you @sajal2692 for letting me know of the inconsistency in #20, and for your work on this repo!

---

- Tom Aarsen